destroyメソッド削除条件追加

+++ app/controllers/items_controller.rb
@@ -27,7 +27,7 @@ class ItemsController < ApplicationController
     unless current_user == User.find(@item.user_id)
       redirect_to action: :index
     end
-    redirect_to action: :index
+
   end
 
   def update
@@ -39,6 +39,9 @@ class ItemsController < ApplicationController
   end
 
   def destroy
+    unless current_user == User.find(@item.user_id)
+      redirect_to action: :index
+    end
     item = Item.find(params[:id])
     item.destroy
     redirect_to root_path
